+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Migrate to AndroidX for managing Fragments
 Allow for sending multiple files at once on Android 21+
 
 **Breaking Change**
+Initialization
+* removed `setConfiguration` method from `ChatWindowView`
+* changed `initialize` to `init` and added required `ChatWindowConfiguration` parameter
+
 Move away from deprecated `startActivityForResult` to `registerForActivityResult` to start file picker Activity and handle its result.
 
 Removed: